requirements.txt
Flask==3.0.3
numpy==1.25.1
pandas==2.0.3
tensorflow==2.14.0
PyPDF2==3.0.1
pdfminer.six==20240706
pypandoc==1.14
python-docx==1.1.2
faiss-cpu==1.9.0
openai==1.57.2
tiktoken==0.8.0
transformers==4.31.0
uvicorn==0.32.0
dearpygui==1.11.1
